+++ src/Rules/ArrayOfRule.php
@@ -2,6 +2,7 @@
 
 namespace Orisai\ObjectMapper\Rules;
 
+use Orisai\Exceptions\Logic\InvalidArgument;
 use Orisai\ObjectMapper\Args\Args;
 use Orisai\ObjectMapper\Args\ArgsChecker;
 use Orisai\ObjectMapper\Context\ArgsFieldContext;
@@ -15,7 +16,9 @@ use Orisai\ObjectMapper\Types\GenericArrayType;
 use Orisai\Utils\Arrays\ArrayMerger;
 use function array_values;
 use function count;
+use function get_debug_type;
 use function is_array;
+use function sprintf;
 
 /**
  * @extends MultiValueRule<ArrayOfArgs>
@@ -61,6 +64,20 @@ final class ArrayOfRule extends MultiValueRule
 		$mergeDefaults = false;
 		if ($checker->hasArg(self::MergeDefaults)) {
 			$mergeDefaults = $checker->checkBool(self::MergeDefaults);
+		}
+
+		if (
+			$mergeDefaults
+			&& $context->hasDefaultValue()
+			&& !is_array($defaultValue = $context->getDefaultValue())
+		) {
+			throw InvalidArgument::create()
+				->withMessage(sprintf(
+					'Argument "%s" given to "%s" is set to true but the default value is "%s" insteadof an array.',
+					self::MergeDefaults,
+					self::class,
+					get_debug_type($defaultValue),
+				));
 		}
 
 		return new ArrayOfArgs(